day_3/homework_1_rag_walkthrough/pyproject.toml
[tool.poetry]
name = "rag-lab"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "1.79.0"
psycopg2-binary = "2.9.10"
python-dotenv = "1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
